# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Personas y Aulas",
        "version": "1.0.0"
    },
    "paths": {
        "/api/aulas": {
            "get": {
                "tags": [
                    "Aulas"
                ],
                "summary": "Listar aulas",
                "operationId": "b26bb8375b61d31b823e3c698cc3ae8f",
                "responses": {
                    "200": {
                        "description": "Lista de aulas"
                    }
                }
            },
            "post": {
                "tags": [
                    "Aulas"
                ],
                "summary": "Crear aula",
                "operationId": "cd6e8064e2fcadb282ccdbc039462f71",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre",
                                    "id_sede",
                                    "codigo",
                                    "capacidad",
                                    "tipo",
                                    "descripcion"
                                ],
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "Juan"
                                    },
                                    "id_sede": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "codigo": {
                                        "type": "integer",
                                        "example": 101
                                    },
                                    "capacidad": {
                                        "type": "integer",
                                        "example": 30
                                    },
                                    "tipo": {
                                        "type": "string",
                                        "example": "aula"
                                    },
                                    "descripcion": {
                                        "type": "string",
                                        "example": "Aula de matemáticas"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Aula creada"
                    }
                }
            }
        },
        "/api/aulas/{id}": {
            "put": {
                "tags": [
                    "Aulas"
                ],
                "summary": "Actualizar aula",
                "operationId": "6b89a6fdcc6702721ec908b6bc9bb64a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la aula",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre",
                                    "id_sede",
                                    "codigo",
                                    "capacidad",
                                    "tipo",
                                    "descripcion"
                                ],
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "Juan"
                                    },
                                    "id_sede": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "codigo": {
                                        "type": "integer",
                                        "example": 101
                                    },
                                    "capacidad": {
                                        "type": "integer",
                                        "example": 30
                                    },
                                    "tipo": {
                                        "type": "string",
                                        "example": "aula"
                                    },
                                    "descripcion": {
                                        "type": "string",
                                        "example": "Aula de matemáticas"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Aula actualizada"
                    },
                    "404": {
                        "description": "Aula no encontrada"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Aulas"
                ],
                "summary": "Eliminar aula",
                "operationId": "8eba2a81113b30099462c4bf1fb17099",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la aula",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Aula eliminada"
                    },
                    "404": {
                        "description": "Aula no encontrada"
                    }
                }
            }
        },
        "/api/ciudades": {
            "get": {
                "tags": [
                    "Ciudades"
                ],
                "summary": "Listar ciudades",
                "operationId": "1d7f410343dc253f9396a53b4cebf853",
                "responses": {
                    "200": {
                        "description": "Lista de ciudades"
                    }
                }
            },
            "post": {
                "tags": [
                    "Ciudades"
                ],
                "summary": "Crear ciudad",
                "operationId": "72e679555bad30853c49755de8505641",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre",
                                    "id_pais"
                                ],
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "Juan"
                                    },
                                    "id_pais": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Ciudad creada"
                    }
                }
            }
        },
        "/api/ciudades/{id}": {
            "put": {
                "tags": [
                    "Ciudades"
                ],
                "summary": "Actualizar ciudad",
                "operationId": "d6e78851d2536a726294c972f8d0a694",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la ciudad",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre",
                                    "id_pais"
                                ],
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "Juan"
                                    },
                                    "id_pais": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ciudad actualizada"
                    },
                    "404": {
                        "description": "Ciudad no encontrada"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Ciudades"
                ],
                "summary": "Eliminar ciudad",
                "operationId": "09476657ede195b4ffbc9750f42d6914",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la ciudad",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Ciudad eliminada"
                    },
                    "404": {
                        "description": "Ciudad no encontrada"
                    }
                }
            }
        },
        "/api/clases": {
            "get": {
                "tags": [
                    "Clases"
                ],
                "summary": "Listar clases",
                "operationId": "6b51d92341941d3b22829adbd68dfa7b",
                "responses": {
                    "200": {
                        "description": "Lista de clases"
                    }
                }
            },
            "post": {
                "tags": [
                    "Clases"
                ],
                "summary": "Crear clase",
                "operationId": "f55bdd808ee3b056f884b4b179a328f5",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre",
                                    "apellido"
                                ],
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "Juan"
                                    },
                                    "apellido": {
                                        "type": "string",
                                        "example": "Pérez"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Clase creada"
                    }
                }
            }
        },
        "/api/clases/{id}": {
            "put": {
                "tags": [
                    "Clases"
                ],
                "summary": "Actualizar clase",
                "operationId": "82a086a91ab3bf2e109e92684fadbb34",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la clase",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre",
                                    "apellido"
                                ],
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "Juan"
                                    },
                                    "apellido": {
                                        "type": "string",
                                        "example": "Pérez"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Clase actualizada"
                    },
                    "404": {
                        "description": "Clase no encontrada"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Clases"
                ],
                "summary": "Eliminar clase",
                "operationId": "bb3f338fc77b4722fb422a2063159a6b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la clase",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Clase eliminada"
                    },
                    "404": {
                        "description": "Clase no encontrada"
                    }
                }
            }
        },
        "/api/clases_estudiantes": {
            "get": {
                "tags": [
                    "Clases Estudiantes"
                ],
                "summary": "Listar clases_estudiantes",
                "operationId": "c1be90931049e9080eba53c6a9a2002c",
                "responses": {
                    "200": {
                        "description": "Lista de clases_estudiantes"
                    }
                }
            },
            "post": {
                "tags": [
                    "Clases Estudiantes"
                ],
                "summary": "Crear clase_estudiante",
                "operationId": "b7e81ebd3589a81cb079a53b6db84b21",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id_clase",
                                    "id_estudiante"
                                ],
                                "properties": {
                                    "id_clase": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "id_estudiante": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Clase_Estudiante creada"
                    }
                }
            }
        },
...
        "/api/clases_horarios": {
            "get": {
                "tags": [
                    "Clases Horarios"
                ],
                "summary": "Listar clases_horarios",
                "operationId": "be078db6aeeca21c12d566271aae3e18",
                "responses": {
                    "200": {
                        "description": "Lista de clases_horarios"
                    }
                }
            },
            "post": {
                "tags": [
                    "Clases Horarios"
                ],
                "summary": "Crear clase_horario",
                "operationId": "8ec1c50abcca438b26ae572d9a8bd734",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id_clase",
                                    "id_horario"
                                ],
                                "properties": {
                                    "id_clase": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "id_horario": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Clase_Horario creada"
                    }
                }
            }
        },
...
    },
...
}